styling added

+++ src/components/feed/Feed.jsx
@@ -1,19 +1,27 @@
 import React, { Component } from 'react';
 
+import classes from './Feed.module.css';
+
 
 import FeedItem from './feedItem/FeedItem';
+import Button, { TYPES } from '../UI/button/Button';
+import Spinner from '../UI/Spinner/Spinner';
+import Modal from '../UI/Modal/Modal';
+import NewPost from '../NewPost/NewPost';
 
-import './Feed.css';
 
 export class Feed extends Component {
   constructor() {
     super();
     this.state = {
-      posts: []
+			posts: [],
+			loading: false,
+			addingPost: false
 		}
 	}
 		
 		componentDidMount() {
+			this.setState({loading: true});
 			this.getPosts();
 		}
 
@@ -25,12 +33,14 @@ export class Feed extends Component {
         Authorization: `Bearer ${this.props.authService.getToken()}`,
         "Content-Type": "application/json"
       }
-    })
-      .then(resp => resp.json())
+		})
+			.then(resp => resp.json())
       .then(resp => {
         this.setState(oldState => ({...oldState, posts: resp.posts}));
 			})
-			
+			.then(resp => {
+				this.setState({loading: false});
+			})
       .catch(e => {
         console.log(e);
       });
@@ -38,7 +48,7 @@ export class Feed extends Component {
 
 	// deleteItem = (id) => {
 	// 	console.log(id);
-	// 	fetch(`${this.props.authService.domain}/feed/posts/${id}`, {
+	// 	fetch(`${this.props.authService.domain}/posts/${id}`, {
 	// 		method: "DELETE",
 	// 		headers: {"Content-Type": "application/json"},
 	// 		body: JSON.stringify(id)
@@ -57,24 +67,47 @@ export class Feed extends Component {
   //   })
 	// }
 
+	addPostCancelHandler = () => {
+		this.setState({addingPost: false});
+	}
+
+	addComment = () => {
+		this.setState({addingPost: true});
+	}
+
 
 	
 
   render() {
-		const listComments = this.state.posts.map((posts, i) => {
-			return <FeedItem 
-				id={posts._id}
-				title={posts.title}
-				content={posts.content}
-				name={posts.creator.name}
-				date={posts.createdAt}
-				key={i}
-				{...this.props}
-				deleteComment={this.deleteItem}
-			/>
-		})
-    return <div>
-			<h1>This is your posts</h1>
+		let listComments = null;
+
+		if (this.state.loading) {
+			listComments = <Spinner />;
+		} else {
+			listComments = this.state.posts.map((posts, i) => {
+				return <FeedItem 
+					id={posts._id}
+					title={posts.title}
+					content={posts.content}
+					name={posts.creator.name}
+					date={posts.createdAt}
+					key={i}
+					{...this.props}
+					deleteComment={this.deleteItem}
+				/>
+			});
+		}
+
+    return <div className={classes.Feed}>
+			<Modal show={this.state.addingPost} modalClosed={this.addPostCancelHandler}>
+				<NewPost/>
+			</Modal>
+			<h3>Posts</h3>
+			<Button 
+				className={classes.btn}
+				title={'ADD COMMENT'} 
+				type={ TYPES.default }
+				onClick={this.addComment}/>
 				{listComments}
 			</div>
